refactoring: introduce api module
Split up the content of `shell.py` in two parts.

- `api.py`: a generic module than can be reused by anyone
- `shell.py`: a CLI that consume `api.py`.
